refactored poor variable names

+++ src/main/java/entities/Order.java
@@ -101,7 +101,7 @@ public class Order {
         String start = "The Information of the Delivery Person completing your order is: \n";
         String delName = this.deliveryMan.getName();
         String delContact = this.deliveryMan.getNumber();
-        float delRate = this.deliveryMan.getRate();
+        double delRate = this.deliveryMan.getRate();
         String displayName = "Name: " + delName + "\n";
         String displayContact = "Phone Number: " + "\n";
         String displayRate = "Rating: "  + "\n";
